Streamlined some of the recent tests around the Base Configuration page and broke them out to be in a separate test file.
